Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -3,7 +3,7 @@ int main() {
     int n1,n2,n3;
     float res;
     scanf("%d%d%d",&n1,&n2,&n3);
-    res = (n1+n2+n3)/3;
+    res = (float)(n1+n2+n3)/3;
     printf("Average: %.2f",res);
     
     return 0;
